Git: remove ref bit from `git tr`
The ref bit has an undefined length and pushes the subject off the end
of the line. I don't like having the lines wrap and I also don't like
truncating and losing some or all of the subject line that should fit in
a terminal width.

The ref bit is especially annoying when there are many branches or when
HEAD has more than one, which is common when there are remote branches.

I'll reconsider how to show this information because I would like to be
able to view it easily without opening gitk.