Vim: Adjust ALE's linting criteria and delay
Having the quickfix window open and close every few hundred milliseconds
while typing is very annoying. This change runs linting only when
leaving insert mode and when text changes in normal mode (in addition to
the defaults of on save and when opening a buffer), instead of any time
text changes (in insert mode).

I've also reduced the delay before linting from the default 200
milliseconds to 50 milliseconds.