Vim: Change ALE's pylint directory
By default, ALE changes the working directory to the file's directory to
run pylint. This is supposed to help it find the correct pylintrc, but
all projects I've worked on are configured to use pylint's default
behaviour for finding the rc file. This commit change's ALE's default
because so that pylintrc can use its default behaviour.