Zsh: Change theme to agnosterzak
I've found that it's very useful to have a datetime and path displayed
in the shell for a variety of reasons including the value of the
information in screenshots during incidents.

The AgnosterZak theme for Zsh has this information and more and displays
it on a separate line from the prompt to keep commands short. I also
like how it looks.

https://github.com/zakaziko99/agnosterzak-ohmyzsh-theme